send alarm critical and sigsegv on sim-session
Bug fixed: Check if the event returned by the function
sim_command_get_event is null and why

Bug fixed: On sim_connect_send_alarm, when the alarm can't be send to
the framework and then we return the event to the queue. The bug is that
we weren't increasing the reference counter of the object.